Refactor Task and Team models, update TaskDto and UserLoginDto, enhance UserController and TaskController, and improve email sending in EmailService

+++ src/main/java/com/progresso/backend/controller/ProjectController.java
@@ -32,8 +32,8 @@ public class ProjectController {
   }
 
   @GetMapping("/{projectId}/completion")
-  public ResponseEntity<Integer> getProjectCompletionPercentage(@PathVariable Long projectId) {
-    int percentage = projectService.getProjectCompletionPercentage(projectId);
+  public ResponseEntity<Long> getProjectCompletionPercentage(@PathVariable Long projectId) {
+    long percentage = projectService.getProjectCompletionPercentage(projectId);
     return ResponseEntity.ok(percentage);
   }
